+++ src/main/java/gregtech/loaders/recipe/handlers/WireRecipeHandler.java
@@ -186,7 +186,7 @@ public class WireRecipeHandler {
                 .buildAndRegister();
     }
 
-    private static int getVoltageMultiplier(Material material) {
+    private static long getVoltageMultiplier(Material material) {
         return material.getBlastTemperature() >= 2800 ? VA[LV] : VA[ULV];
     }
 }
